Chore: Change the second example of the prompt
The order of lying signals in previous second example was exactly same as the first example's one.
So the script generated by the GPT model also had the same order of lying signals.
To prevent this, a new example was created in which the order of lying signals was changed.